AP-7137: Add nodes to binary features
The license for the tableau binary states that it shouldn't be
distributed without added content. Therefore I just added the plugins
containing the nodes to the feature of the binaries. People who still
don't want the binaries from the update site can use the other feature
but only installing the binaries without nodes doesn't make sense.