AP-11459: Check exists only on append
The Tableau REST API acts like expected if overwrite is true:
Overwrite if exists, create if doesn't exist
Only on append the REST API acts differently: Fail if doesn't exist.
Therefore the check if the datasource exists is only needed for append
and the overwrite flag can be set always if the user selected the option.